pip_requirements.txt
absl-py==0.8.1
algorithmia==1.2.1
algorithmia-api-client==1.0.1
anyio==3.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.8.1
bitstruct==8.15.1
bleach==3.3.0
bleak==0.15.1
bs4==0.0.1
CacheControl==0.12.6
cachetools==3.1.1
certifi==2019.9.11
cffi==1.14.5
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.1
CodernityDB==0.5.0
colorama==0.4.1
coverage==5.3.1
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
drawnow==0.72.0
entrypoints==0.3
enum34==1.1.6
Faker==5.0.1
ffmpeg==1.4
Fiona==1.8.18
firebase-admin==4.5.0
Flask==1.1.1
gast==0.3.3
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.22.0
google-images-download==2.8.0
google-pasta==0.2.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
googlemaps==4.4.2
grpcio==1.24.1
gTTS==2.0.4
gTTS-token==1.1.3
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.6.1
importlib-metadata==3.3.0
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Lasagne==0.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
matplotlib-inline==0.1.2
micropyGPS==1.0
mistune==0.8.4
mongoengine==0.18.2
msgpack==0.6.2
munch==2.5.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
playsound==1.2.2
plotly==4.14.3
plotly-express==0.4.1
pluggy==0.13.1
pool==0.1.2.dev0
pprintpp==0.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.10.0
ptyprocess==0.7.0
py==1.10.0
py-espeak-ng==0.1.8
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycountry==20.7.3
pycountry-convert==0.7.2
pycparser==2.20
pydub==0.23.1
pygame==2.0.0
Pygments==2.9.0
pymongo==3.9.0
pynput==1.4.5
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-Quartz==6.1
pyparsing==2.4.5
pyproj==3.0.0.post1
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
pytest==6.2.1
pytest-cov==2.10.1
pytest-mock==3.4.0
python-dateutil==2.8.0
pyttsx==1.1
pytz==2019.3
pyzmq==22.1.0
repoze.lru==0.7
requests==2.22.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
retrying==1.3.3
reverse-geocoder==1.5.1
rsa==4.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
simplekml==1.3.5
six==1.12.0
sniffio==1.2.0
soupsieve==1.9.4
statsmodels==0.12.1
stem==1.7.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
Theano==1.0.4
tinydb==3.15.1
toml==0.10.2
tornado==6.1
torrequest==0.1.0
tqdm==4.54.1
traitlets==5.0.5
typing_extensions==4.3.0
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.4.0
